chore: Migrate inline embedded content
Whilst working on NP-1735 (migrating anchor links) it became apparent
that we weren't handling inline embedded entries properly during the
conversion stage, so it was causing validation issues.  Essentially we
need to keep inline embedded entries within the paragraph they are in to
not only maintain the original order, but also to pass the Contentful
validation checks as they're expecting inline entries to be within a
paragraph nodetype.

I've also renamed the feature spec as realised this wasn't being
automatically picked up/run by Rspec because the naming convention was
wrong.