Stop creating entries for internal links
One of the biggest issues we have during the migration is when trying to
publish content and getting the error "linked entry is missing or
inaccessible".  When we hit the Contentful converter it would create an
entry, rather than just a plain hyperlink.  This means when we're trying
to publish a page Contentful is looking for an entry which doesn't
actually exist, and prevents us from publishing.

We can also end up with empty links and this too was creating an
entry-link with no ID.  I think it would be better to just leave these
as hyperlinks to avoid Contentful looking for an entry without an ID.