Stop adding a space to ul items
During the Family section sense check a number of issues were logged in
regards to extra spaces being added to some links/sentences.  It turns
out that we have been adding an extra space when creating ul list items
which resulted in things like "savings disregard .".  I've removed the
step that adds the extra space, done a migration to QA with this change
and Claire has been through the content fix sheet to check them all.
According to her checks this has resolved all these issues.

It wasn't clear why we were adding this space, and through the checks
that both Claire and I have done we haven't come across any issues thus
far that would indicate that this change is causing any knock on
effects, but this is something that we can keep an eye on and review
again if needs be